+++ NEWS.md
@@ -8,6 +8,7 @@
 * Add constants for error codes in C API. The following macro are defined:
   `CHFL_SUCCESS`, `CHFL_MEMORY_ERROR`, `CHFL_FILE_ERROR`, `CHFL_FORMAT_ERROR`,
   `CHFL_GENERIC_ERROR`, `CHFL_CXX_ERROR`.
+* Add the `chfl_version` function in C API.
 
 # 0.4